requirements.txt
fastapi
uvicorn
python-dotenv
jinja2
sqlalchemy
openai>=1.0.0
python-multipart
email-validator